Rename TalkerHistoryBuilder -> TalkerBuilder and remove old TalkerBuilder
