+++ packages/talker_flutter/lib/src/src.dart
@@ -1,6 +1,7 @@
 export 'extensions/extensions.dart';
 export 'observers/observers.dart';
 export 'ui/builders/builders.dart';
+export 'ui/talker_builder.dart';
 export 'ui/talker_listener.dart';
 export 'ui/talker_screen.dart';
 export 'ui/talker_wrapper/talker_wrapper.dart';
